Refactor and enhance tests for domain and infrastructure components
- Added comprehensive tests for the Distance structure and its entity inheritance.
- Consolidated distance-related tests into a single file, removing the old distances_test.go.
- Removed outdated tests for Entity, MarketHall, and other models, replacing them with enhanced versions.
- Updated HTTP handler tests to improve coverage and ensure proper response formats and error handling.
- Removed old repository tests that required integration setup, marking them as skipped.
- Ensured all tests validate expected behaviors, including consistency, symmetry, and completeness of distance data.
