--- main_enhanced_test.go
@@ -1,189 +1,0 @@
-package main
-
-import (
-	"context"
-	"os"
-	"testing"
-
-	"github.com/adnvilla/patrician/src/domain"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestApplicationSetup(t *testing.T) {
-	t.Run("CitiesInitialization", func(t *testing.T) {
-		// Test that cities are properly initialized as in main
-		for _, city := range domain.Cities {
-			commodities := domain.GetCommodities()
-			city.SetMarketHall(domain.MarketHall{Commodities: commodities})
-		}
-
-		// Verify all cities have market halls with commodities
-		for cityName, city := range domain.Cities {
-			assert.NotNil(t, city.MarketHall, "City %s should have a market hall", cityName)
-			assert.NotNil(t, city.MarketHall.Commodities, "City %s should have commodities", cityName)
-			assert.Equal(t, len(domain.GetCommodities()), len(city.MarketHall.Commodities),
-				"City %s should have all commodities", cityName)
-		}
-	})
-
-	t.Run("DomainConsistency", func(t *testing.T) {
-		// Test that domain objects are consistent
-		cities := domain.Cities
-		commodities := domain.GetCommodities()
-		distances := domain.Distances
-
-		assert.Greater(t, len(cities), 0, "Should have cities")
-		assert.Greater(t, len(commodities), 0, "Should have commodities")
-		assert.Greater(t, len(distances), 0, "Should have distances")
-
-		// Cities and distances should be consistent
-		for cityName := range cities {
-			_, hasDistances := distances[cityName]
-			assert.True(t, hasDistances, "City %s should have distance entries", cityName)
-		}
-	})
-}
-
-func TestEnvironmentHandling(t *testing.T) {
-	t.Run("DefaultPort", func(t *testing.T) {
-		// Test that the application can handle default port
-		// This is more of a structural test since we can't actually run the server
-
-		// Ensure PORT environment variable behavior
-		originalPort := os.Getenv("PORT")
-		defer os.Setenv("PORT", originalPort)
-
-		// Test with no PORT set
-		os.Unsetenv("PORT")
-		// Application should use default :8080 (this is Gin's default behavior)
-
-		// Test with PORT set
-		os.Setenv("PORT", "9090")
-		// Application should use :9090
-	})
-}
-
-func TestOpenAPISetup(t *testing.T) {
-	t.Run("OpenAPIFileExists", func(t *testing.T) {
-		// Test that the OpenAPI file exists
-		_, err := os.Stat("docs/openapi.yaml")
-		if err != nil {
-			// File might not exist in test environment, that's okay
-			t.Skip("OpenAPI file not found, skipping test")
-		}
-	})
-
-	t.Run("ContextUsage", func(t *testing.T) {
-		// Test that context is used properly
-		ctx := context.Background()
-		assert.NotNil(t, ctx)
-
-		// Test context timeout (example of how context might be used)
-		_, cancel := context.WithCancel(ctx)
-		defer cancel()
-	})
-}
-
-func TestRouterConfiguration(t *testing.T) {
-	t.Run("ExpectedRoutes", func(t *testing.T) {
-		// Test that we have the expected routes defined
-		// This is more of a documentation test
-
-		expectedRoutes := map[string]string{
-			"GET /cities":                  "GetCities",
-			"GET /commodities":             "GetCommodities",
-			"GET /distances":               "GetDistances",
-			"GET /city/:name/commodities":  "GetCityCommodities",
-			"POST /city/:name/commodity":   "UpdateCommodity",
-			"POST /city/:name/commodities": "UpdateCommodities",
-			"GET /city/:name/stock":        "GetStock",
-			"GET /city/:name/supply/:city": "GetSupply",
-		}
-
-		// We can't easily test the actual router without running it,
-		// but we can verify the handlers exist
-		assert.NotNil(t, expectedRoutes)
-		assert.Equal(t, 8, len(expectedRoutes))
-	})
-}
-
-func TestApplicationDependencies(t *testing.T) {
-	t.Run("RequiredPackages", func(t *testing.T) {
-		// Test that required packages are importable
-		// This is a compile-time test essentially
-
-		// Domain package
-		cities := domain.Cities
-		assert.NotNil(t, cities)
-
-		commodities := domain.GetCommodities()
-		assert.NotNil(t, commodities)
-
-		distances := domain.Distances
-		assert.NotNil(t, distances)
-	})
-}
-
-func TestApplicationConstants(t *testing.T) {
-	t.Run("DomainConstants", func(t *testing.T) {
-		// Test that domain constants are available
-		assert.Equal(t, float32(30), domain.Day)
-		assert.Equal(t, float32(210), domain.Week)
-		assert.Equal(t, domain.CommodityType(1), domain.Load)
-		assert.Equal(t, domain.CommodityType(2), domain.Barrel)
-	})
-}
-
-func TestDataIntegrity(t *testing.T) {
-	t.Run("CitiesCommoditiesConsistency", func(t *testing.T) {
-		// Test the data setup as done in main()
-		for _, city := range domain.Cities {
-			commodities := domain.GetCommodities()
-			city.SetMarketHall(domain.MarketHall{Commodities: commodities})
-		}
-
-		// All cities should have the same commodities available
-		expectedCommodityCount := len(domain.GetCommodities())
-
-		for cityName, city := range domain.Cities {
-			actualCommodityCount := len(city.MarketHall.Commodities)
-			assert.Equal(t, expectedCommodityCount, actualCommodityCount,
-				"City %s should have %d commodities", cityName, expectedCommodityCount)
-		}
-	})
-
-	t.Run("CommodityTypesConsistency", func(t *testing.T) {
-		commodities := domain.GetCommodities()
-
-		for name, commodity := range commodities {
-			// Each commodity should have a valid type
-			assert.Contains(t, []domain.CommodityType{domain.Barrel, domain.Load},
-				commodity.CommodityType, "Commodity %s should have valid type", name)
-		}
-	})
-}
-
-// TestMainFunction tests aspects of the main function without actually running it
-func TestMainFunction(t *testing.T) {
-	t.Run("MainFunctionSetup", func(t *testing.T) {
-		// We can't easily test main() directly, but we can test the setup logic
-
-		// Test city initialization logic
-		originalCities := make(map[string]*domain.City)
-		for name, city := range domain.Cities {
-			originalCities[name] = city
-		}
-
-		// Apply the same logic as in main()
-		for _, city := range domain.Cities {
-			commodities := domain.GetCommodities()
-			city.SetMarketHall(domain.MarketHall{Commodities: commodities})
-		}
-
-		// Verify setup was successful
-		for cityName, city := range domain.Cities {
-			assert.NotNil(t, city.MarketHall, "City %s should have market hall after setup", cityName)
-			assert.NotNil(t, city.MarketHall.Commodities, "City %s should have commodities after setup", cityName)
-		}
-	})
-}
