--- src/infrastructure/data/postgresql/cities_repository_test.go.bak
@@ -1,10 +1,0 @@
-package postgresql_test
-
-import (
-	"testing"
-)
-
-// Skipping this test suite for now as it requires more integration
-func TestCitiesRepository(t *testing.T) {
-	t.Skip("Skipping cities repository tests as they require additional integration setup")
-}
