--- src/infrastructure/data/postgresql/commodities_repository_test.go.bak
@@ -1,10 +1,0 @@
-package postgresql_test
-
-import (
-	"testing"
-)
-
-// Skipping this test suite for now as it requires more integration
-func TestCommoditiesRepository(t *testing.T) {
-	t.Skip("Skipping commodities repository tests as they require additional integration setup")
-}
