--- src/infrastructure/data/postgresql/distances_repository_test.go.bak
@@ -1,31 +1,0 @@
-package postgresql_test
-
-import (
-	"testing"
-)
-
-// Skipping this test suite for now as it requires more integration
-func TestDistancesRepository(t *testing.T) {
-	t.Skip("Skipping distances repository tests as they require additional integration setup")
-}
-		
-		// Assert
-		assert.Error(t, err)
-		assert.Nil(t, distance)
-	})
-}
-		
-		// Assert
-		assert.Error(t, err)
-		assert.Nil(t, distance)
-	})
-}
-
-		// Act
-		distance, err := repo.GetDistanceBetweenCities(ctx, fromCity, toCity)
-
-		// Assert
-		assert.Error(t, err)
-		assert.Nil(t, distance)
-	})
-}
