--- src/infrastructure/data/postgresql/markethall_repository_test.go.bak
@@ -1,10 +1,0 @@
-package postgresql_test
-
-import (
-	"testing"
-)
-
-// Skipping this test suite for now as it requires more integration
-func TestMarketHallRepository(t *testing.T) {
-	t.Skip("Skipping markethall repository tests as they require additional integration setup")
-}
